requirements.txt
# Core Dependencies
polars>=0.20.0  # High-performance DataFrames (Staff choice over Pandas)
duckdb>=0.9.0  # OLAP database (The local Snowflake)
requests>=2.31.0 # HTTP Client
pyyaml>=6.0.1  # Config management
pydantic>=2.0.0 # Type validation (Staff practice)
retrying>=1.3.3 # Exponential backoff

# Visualization
streamlit>=1.28.0
plotly>=5.17.0

# Engineering
pytest>=7.4.0
python-dotenv>=1.0.0
loguru>=0.7.0 # Advanced logging

# Database (Supabase/Postgres)
adbc-driver-postgresql>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
